make customFetch global (#386)
To try and reduce the likelihood of this breaking again we can simply register customFetch globally and use it directly from the window. It still relies on wasm-bindgen not changing a specific line, but at least it's one line instead of 5.

I tested it locally and it works.

Obviously doing things globally isn't ideal, but I renamed it to be a bit more explicit.

I tried simply overriding `fetch` itself globally, but it broke everything.