add intro text to assets page (#404)
## Objective

The assets page has no text describing what it actually is, it should be made a bit clearer. 

## Solution
I used the description from the bevy_assets README to clarify it a bit more.

![vivaldi_znFCzIkiSO](https://user-images.githubusercontent.com/8348954/180702648-853d7a8b-481a-45f6-b1d4-8c5ff5342d06.png)

## Notes

I'm sure some people would like a clear disclaimer related to licenses. I'm open to the idea, but this could be done in a separate PR and shouldn't block this PR.

Co-authored-by: Charles <IceSentry@users.noreply.github.com>