add "before migrating" section to migration guide (#428)
## Objective

It's not uncommon for people to forget to update their rust versions. 

## Solution

Add a short warning to remind people to run `rustup update` and a short description as to why

![image](https://user-images.githubusercontent.com/8348954/182221263-3f079b08-22e6-40c9-9d09-9296b75d4490.png)

## Notes

For now this is only on the latest migration guide, but I already have plans to make migration guide a dedicated section on the website and use a template that could include this warning on every page.