Add Mold to "recommended linker" section (#442)
> mold is a faster drop-in replacement for existing Unix linkers. It is several times faster than the LLVM lld linker, the second-fastest open-source linker which I originally created a few years ago. mold is designed to increase developer productivity by reducing build time, especially in rapid debug-edit-rebuild cycles.

| Program (linker output size)  | GNU gold | LLVM lld | mold
|-------------------------------|----------|----------|--------
| Chrome 96 (1.89 GiB)          | 53.86s   | 11.74s   | 2.21s
| Clang 13 (3.18 GiB)           | 64.12s   | 5.82s    | 2.90s
| Firefox 89 libxul (1.64 GiB)  | 32.95s   | 6.80s    | 1.42s

Mold is exclusive to Linux, with support for Windows and MacOS in progress.

---

This PR adds mold to the recommended linkers in [the setup page](https://bevyengine.org/learn/book/getting-started/setup/).

Note that this addition is somewhat clunky, and this section may need to be reworked entirely. See #329 for further discussion.