Generate extra metadata from external sources (#397)
## Objective

The new asset cards (#394) need more metadata for license and bevy compatible versions.

## Solution

- For assets with a link to github or gitlab:
  - Get the Cargo.toml file using the apis, parse it and read the bevy version and license fields.
  - This is done on a best effort basis. It's not always possible to have this information. For example, when the Cargo.toml is not at the top level. 
- For assets with crates.io links
  - Use the crates.io dbdump to get the required data.
- If the metadata is entered manually directly in bevy_assets it will be used instead.

The github client relies on a GITHUB_TOKEN env var to exist in CI or locally. It should be already there since we use github actions. The rate limit in github actions is 1000 per hour https://docs.github.com/en/rest/overview/resources-in-the-rest-api#requests-from-github-actions so we should be fine for a little while.

The gitlab client implementation ignores the token because we have so few crates that it doesn't matter.

## Notes

Querying the crates.io dump is pretty slow and the deps-tree required is pretty big. I think we should consider eventually rolling our own way to do that, but it works well right now.

This is going to be pretty slow to execute in the job once it is merged, but it's pretty easy to parallelize all the generate-* jobs so I did it in #401 .