Fix broken assets page (#445)
Fixes #444

It's not completely clear to me exactly what was intended with this change in #401, but this logic seems obviously reversed (it just completely erased the path).

This fixes the assets page locally for me.

edit: I see the note in 401 now about the paths needing to be filtered to make some part of the the CI / zipping work. Not entirely sure how to test that.