News: Release 0.9.0 (#478)
This is an incomplete draft of the 0.9.0 release blog post. Feel free to leave comments / suggest changes / help me fill in gaps I missed!

NOTE: This is an _early_ draft and there are still a lot of missing pieces / placeholders. Please only review content that exists and point out missing pieces that don't already have placeholders. Anything with a `[4724]`-style markdown link is a placeholder for content that _might_  need to be included (but also might not be ... these placeholders still need some curation).