Community: Add Role tags and consolidate Bevy Organization (#514)
This adds Bevy Organization Role "tags" to the Bevy People/Community page (Project Lead, Maintainer, SME). These are displayed on each card. I've also added a short description of the Bevy Org and each role.
It also consolidates all org sections (Maintainers, Bevy Org) into a single "The Bevy Organization" section. 

You might notice we have a new role! We'll be announcing what that is and the people that will occupy it in the next day or so.

The Bevy People page now looks like this.

![image](https://user-images.githubusercontent.com/2694663/212442625-b3f90ed6-586e-444f-a84d-b2767e1ac91c.png)

Note that this currently uses my fork of bevy-community. We should merge this first, then migrate bevy-community to the new format.