Make the People page less confusing (#518)
When I first looked at this page, I found it confusing. I had to read the README in the `bevy-community` repo to understand how it all works and why there is a separate Community Members section.

My first thoughts were "why are there so many people with no Role tags under the bevy org section? what is their role in the org? seems like there are a bunch of random people listed here, and some are separated under the community heading and others lumped in with the project people?"

This PR helps make the page more understandable.
 
 - Add some text under "Roles", to describe the role of the org members without any "special role" tag
 - Add some text under "Community Members", to clarify the purpose of that section