feat: use `create_dir_all`
`create_dir_all` is the infallible variant of `create_dir` that creates both the specified folder and all of its parents, if they are missing.

I'm not sure this line is even necessary any more. Is there ever a circumstance where the content directory does not exist? If show, shouldn't it be an error?