Hard-code possible docs root sections
The previous, more procedural, way of determing docs root sections did not take into consideration possible docs sections outside of `/learn/*`. I would've done a better procedural way, but the lack of while loops in Tera threw a wrench in that plan, so for now this is the best I can do without further blocking progress on the contributing guide.