Change initial reference to Bevy to the Bevy Organization
Technically the Bevy Organization projects you can contribute to expands past the game engine, primarily the website, so I feel it is a good idea to make this distinction since a game engine is wildly different to a website for what you can contribute to it.

Also maybe some substantive editing regarding difference of Organization and Foundation may be needed to clarify things.