Improve the release notes generator
It now tracks whether a PR exists by the PRs metadata, and allows for mutliple PRs per a note, and PR URLs per note. It should only ever create a file if a PR is not included in any notes metadata, so no more overwriting note files or note metadata if it isn't a new PR as far as the generator is concerned.